+++ src/cmd/inventory.rs
@@ -10,7 +10,7 @@ use crate::{
     inventory::{default_inventory_path, models::Inventory},
     io::{self, encoding::InventoryPostProcessor, Scanner},
 };
-use anyhow::{Context, Result};
+use anyhow::{anyhow, bail, ensure, Context, Result};
 use clap::{Args, Subcommand};
 use std::{
     collections::HashSet,
@@ -100,16 +100,16 @@ async fn build(opts: &ConfigOpts) -> Result<()> {
         .verify
         .as_ref()
         .and_then(|opts| opts.token.clone())
-        .context("no token was provided")?;
+        .context("no token was provided, need help? see https://github.com/simbleau/cddns#readme")?;
 
     // Get zones and records to build inventory from
     info!("retrieving cloudflare resources...");
     let mut zones = cloudflare::endpoints::zones(&token).await?;
     crate::cmd::list::retain_zones(&mut zones, opts)?;
-    anyhow::ensure!(!zones.is_empty(), "no zones to build inventory from");
+    ensure!(!zones.is_empty(), "no zones to build inventory from");
     let mut records = cloudflare::endpoints::records(&zones, &token).await?;
     crate::cmd::list::retain_records(&mut records, opts)?;
-    anyhow::ensure!(!records.is_empty(), "no records to build inventory from");
+    ensure!(!records.is_empty(), "no records to build inventory from");
 
     let runtime = tokio::runtime::Handle::current();
     let mut scanner = Scanner::new(runtime);
@@ -241,7 +241,7 @@ async fn check(opts: &ConfigOpts) -> Result<()> {
         .verify
         .as_ref()
         .and_then(|opts| opts.token.clone())
-        .context("no token was provided")?;
+        .context("no token was provided, need help? see https://github.com/simbleau/cddns#readme")?;
 
     // Get inventory
     info!("reading inventory...");
@@ -306,7 +306,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
         .verify
         .as_ref()
         .and_then(|opts| opts.token.clone())
-        .context("no token was provided")?;
+        .context("no token was provided, need help? see https://github.com/simbleau/cddns#readme")?;
 
     // Get inventory
     info!("reading inventory...");
@@ -383,7 +383,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
                             )
                             .await
                         }
-                        None => Err(anyhow::anyhow!("no discovered ipv4 address needed to patch A record")),
+                        None => Err(anyhow!("no discovered ipv4 address needed to patch A record")),
                     },
                     "AAAA" => match ipv6 {
                         Some(ip) => update_record(
@@ -393,7 +393,7 @@ async fn commit(opts: &ConfigOpts) -> Result<()> {
                                 ip,
                             )
                             .await,
-                        None => Err(anyhow::anyhow!("no discovered ipv6 address needed to patch AAAA record")),
+                        None => Err(anyhow!("no discovered ipv6 address needed to patch AAAA record")),
                     },
                     _ => unimplemented!(
                             "unexpected record type: {}",
@@ -548,7 +548,7 @@ pub async fn check_records(
                             mismatches.push(cf_record.clone());
                         }
                     } else {
-                        anyhow::bail!(
+                        bail!(
                             "no address comparable for {} record",
                             cf_record.record_type
                         );